+++ endpoint/onramp/controllers/src/test/java/com/hotels/road/onramp/controller/OnrampAuthorisationTest.java
@@ -69,7 +69,6 @@ public class OnrampAuthorisationTest {
 
   private @Mock Counter counter;
 
-  @SuppressWarnings("unchecked")
   @Before
   public void before() throws Exception {
     road.setAuthorisation(authorisation);